Add column delete_original in media_file. This will allow if 0, to not delete the media file where it is stored when row is deleted

+++ library/Centurion/Contrib/media/models/DbTable/Row/File.php
@@ -108,8 +108,10 @@ class Media_Model_DbTable_Row_File extends Centurion_Db_Table_Row_Abstract
 
     public function delete()
     {
-        if ($this->file_id !== null && $this->getTable()->select(true)->where('file_id=?', $this->file_id)->count() == 1) {
-            unlink(Centurion_Config_Manager::get('media.uploads_dir') . DIRECTORY_SEPARATOR . $this->local_filename);
+        if ($this->delete_original == 1) {
+            if ($this->file_id !== null && $this->getTable()->select(true)->where('file_id=?', $this->file_id)->count() == 1) {
+                unlink(Centurion_Config_Manager::get('media.uploads_dir') . DIRECTORY_SEPARATOR . $this->local_filename);
+            }
         }
 
         parent::delete();
@@ -117,8 +119,9 @@ class Media_Model_DbTable_Row_File extends Centurion_Db_Table_Row_Abstract
 
     public function getRelativePath($effects = null, $extra = false, $realPath = false)
     {
-        if (is_array($effects))
+        if (is_array($effects)) {
             $effects = Media_Model_DbTable_Image::effectsArray2String($effects);
+        }
 
         return Centurion_Inflector::urlEncode(pack("H*" , $this->file_id)) . '/_' . ((null !== $effects) ? $effects:'') . '.centurion';
     }
